Maintenance tempalte
app.js is the server with the get function and maintenance.ejs that is the template which gets data from the server and displays it(front end). Need to connect it to the back end. It does not let me upload the node modules I used as it exceed the amount of files I can upload, without that it can't be run properly. If the code needs to be executed then, 'npm install express body-parser' should load the modules up to your local computer , after that just node app.js to run the file and see the template in localhost3000